Split options
